Add working react config with hot reload
