+++ extensions/theme-extension/blocks/chatbot.liquid
@@ -1,169 +1,17 @@
-{% comment %} <script src="https://ecomrag--development.gadget.app/api/client/web.min.js" defer="defer"></script>
-<script src="{{ 'purify.min.js' | asset_url }}" defer="defer"></script>
-<script src="{{ 'chatbot.js' | asset_url }}" defer="defer"></script>
-<script>
-  var shopId = {{ shop.id | json }};
-</script>
-<link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Sarabun:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800&display=swap" rel="stylesheet">
-
-{{ 'chatbot.css' | asset_url | stylesheet_tag }}
-
-<div id="chatbot-toggle" style="bottom: {{ block.settings.toggle_bottom }}; top: {{ block.settings.toggle_top }}; right: {{ block.settings.toggle_right }}; left: {{ block.settings.toggle_left }}">
-  <img
-    id="chatbot-open-toggle"
-    src="{{ "assistant.jpg" | asset_img_url: 'large' }}"
-    height="50"
-    width="50"
-    loading="lazy">
-  <img
-    id="chatbot-close-toggle"
-    class="hidden"
-    src="{{ "close.png" | asset_img_url: 'large' }}"
-    height="35"
-    width="35"
-    loading="eager">
+{% liquid
+  # Relative to entrypointsDir
+  render 'vite-tag' with 'theme.css'
+  render 'vite-tag' with 'theme.tsx'
+%}
+<div 
+  id="root" 
+  style="position: absolute; bottom: 20px; right: 20px;"
+>
 </div>
-<div id="chatbot-window">
-  <div class="chat-header">
-    <div class="chat-header-content">
-
-      <div class="chat-header-avatar-container">
-
-        <img
-          src="{{ 'assistant.jpg' | asset_img_url: 'large' }}"
-          style="border-radius: 50%;"
-          height="40"
-          width="40">
-        <p class="chat-header-title" id="chat-header-title" ></p>
-      </div>
-
-      <div id="close-icon">
-        <svg
-          fill="#fff"
-          height="16px"
-          width="16px"
-          version="1.1"
-          id="Capa_1"
-          xmlns="http://www.w3.org/2000/svg"
-          xmlns:xlink="http://www.w3.org/1999/xlink"
-          viewBox="0 0 460.775 460.775"
-          xml:space="preserve">
-          <path d="M285.08,230.397L456.218,59.27c6.076-6.077,6.076-15.911,0-21.986L423.511,4.565c-2.913-2.911-6.866-4.55-10.992-4.55
-     c-4.127,0-8.08,1.639-10.993,4.55l-171.138,171.14L59.25,4.565c-2.913-2.911-6.866-4.55-10.993-4.55
-     c-4.126,0-8.08,1.639-10.992,4.55L4.558,37.284c-6.077,6.075-6.077,15.909,0,21.986l171.138,171.128L4.575,401.505
-     c-6.074,6.077-6.074,15.911,0,21.986l32.709,32.719c2.911,2.911,6.865,4.55,10.992,4.55c4.127,0,8.08-1.639,10.994-4.55
-     l171.117-171.12l171.118,171.12c2.913,2.911,6.866,4.55,10.993,4.55c4.128,0,8.081-1.639,10.992-4.55l32.709-32.719
-     c6.074-6.075,6.074-15.909,0-21.986L285.08,230.397z" />
-        </svg>
-      </div>
-
-    </div>
-  </div>
-
-  <div id="chat">
-    <p id="introduction-text">
-      {{ "chatbot.welcome.label" | t }}
-    </p>
-  </div>
-  <svg 
-    id="lucide-x"
-    xmlns="http://www.w3.org/2000/svg" 
-    width="24" 
-    height="24" 
-    viewBox="0 0 24 24" 
-    fill="none" 
-    stroke="currentColor" 
-    stroke-width="2" 
-    stroke-linecap="round" 
-    stroke-linejoin="round" 
-  >
-    <path d="M18 6 6 18"/>
-    <path d="m6 6 12 12"/>
-  </svg>
-  <form id="chat-form">
-    <div id="upload-preview-container">
-    </div>
-    <label for="file-upload" class="custom-file-upload">
-      <input type="file" id="file-upload" accept="image/*"/>
-      <svg 
-        xmlns="http://www.w3.org/2000/svg" 
-        width="20" 
-        height="20" 
-        viewBox="0 0 24 24" 
-        fill="none" 
-        stroke="currentColor" 
-        stroke-width="2" 
-        stroke-linecap="round" 
-        stroke-linejoin="round" 
-        class="lucide lucide-upload">
-        <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/><polyline points="17 8 12 3 7 8"/>
-        <line x1="12" x2="12" y1="3" y2="15"/>
-      </svg>
-    </label>
-    <input name="chatInput" id="chat-input" autocomplete="off" />
-    <button name="chatButton" type="submit">
-      <svg
-        fill="#000000"
-        version="1.1"
-        id="Capa_1"
-        xmlns="http://www.w3.org/2000/svg"
-        xmlns:xlink="http://www.w3.org/1999/xlink"
-        width="20px"
-        height="20px"
-        viewBox="0 0 31.806 31.806"
-        xml:space="preserve">
-        <g>
-        <g>
-          <path d="M1.286,12.465c-0.685,0.263-1.171,0.879-1.268,1.606c-0.096,0.728,0.213,1.449,0.806,1.88l6.492,4.724L30.374,2.534
-       L9.985,22.621l8.875,6.458c0.564,0.41,1.293,0.533,1.964,0.33c0.67-0.204,1.204-0.713,1.444-1.368l9.494-25.986
-       c0.096-0.264,0.028-0.559-0.172-0.756c-0.199-0.197-0.494-0.259-0.758-0.158L1.286,12.465z" />
-          <path d="M5.774,22.246l0.055,0.301l1.26,6.889c0.094,0.512,0.436,0.941,0.912,1.148c0.476,0.206,1.025,0.162,1.461-0.119
-       c1.755-1.132,4.047-2.634,3.985-2.722L5.774,22.246z" />
-        </g>
-      </g>
-    </svg>
-  </button>
-</form></div> {% endcomment %}
-<div id="root">
-  {% liquid
-    # Relative to entrypointsDir
-    render 'vite-tag' with 'theme.css'
-    render 'vite-tag' with 'theme.tsx'
-  %}
-</div>
-
-<script>
-  const home = {{ request.path | json }}
-  var shopId = {{ shop.id | json }};
-</script>
 
 {% schema %}
 {
-  "name": "Gadget chatbot",
+  "name": "Gadget Chatbot",
   "target": "body",
-  "settings": [
-    {
-      "type": "text",
-      "label": "bottom",
-      "id": "toggle_bottom",
-      "default": "10px"
-    }, {
-      "type": "text",
-      "label": "top",
-      "id": "toggle_top",
-      "default": "auto"
-    }, {
-      "type": "text",
-      "label": "left",
-      "id": "toggle_left",
-      "default": "auto"
-    }, {
-      "type": "text",
-      "label": "right",
-      "id": "toggle_right",
-      "default": "10px"
-    }
-  ]
+  "settings": []
 }{% endschema %}
