+++ frontend/components/App.tsx
@@ -1,17 +1,12 @@
-import { useState } from 'react'
-import { X } from 'lucide-react';
+import React from 'react';
+import { Toggle } from './buttons/Toggle';
 
-function App() {
-  const [open, setIsOpen] = useState(false);
-
+const App: React.FC<any> = ({ props: _props }) => {
   return (
-    <>
-      <div className='h-4 w-4 rounded-full' onClick={() => setIsOpen((_open) => !_open)}>
-        <img src='assets/assistant.jpg' className={open? 'hidden' : ''} />
-        <X className={open? '' : 'hidden'} />
-      </div>
-    </>
-  )
-}
+    <div className='fixed bottom-5 right-5 z-[1000] flex'>
+      <Toggle />
+    </div>
+  );
+};
 
-export default App
+export default App;
