+++ frontend/entrypoints/vite-env.d.ts
